Modify all the changes that needs to improve
