+++ MVC_Activity01/MVC_Activity01/App_Start/RouteConfig.cs
@@ -16,7 +16,7 @@ namespace MVC_Activity01
             routes.MapRoute(
                 name: "Default",
                 url: "{controller}/{action}/{id}",
-                defaults: new { controller = "Home", action = "About", id = UrlParameter.Optional }
+                defaults: new { controller = "Home", action = "Index", id = UrlParameter.Optional }
             );
         }
     }
