--- MVC_Activity01/MVC_Activity01/Data/MockDb.cs
@@ -1,69 +1,0 @@
-﻿using MVC_Activity01.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace MVC_Activity01.Data
-{
-    public static class MockDb
-    {
-        public static List<Order> Orders = new List<Order>
-        {
-            //new Order
-            //{
-            //    Id = 1,
-            //    No = "0001",
-            //    OrderName = "Computer",
-            //    DocDate = DateTime.Today,
-            //    CustomerName = "Juan",
-            //    Amount = 20000
-            //},
-            //new Order
-            //{
-            //    Id = 2,
-            //    No = "0002",
-            //    OrderName = "AirCon",
-            //    DocDate = DateTime.Today,
-            //    CustomerName = "Pedro",
-            //    Amount = 8500
-            //},
-            // new Order
-            //{
-            //    Id = 3,
-            //    No = "0003",
-            //    OrderName = "Computer CPU",
-            //    DocDate = DateTime.Today,
-            //    CustomerName = "Tomas",
-            //    Amount = 5500
-            //},
-            //  new Order
-            //{
-            //    Id = 4,
-            //    No = "0004",
-            //    OrderName = "Keyboard & Mouse",
-            //    DocDate = DateTime.Today,
-            //    CustomerName = "Carding",
-            //    Amount = 1500
-            //},
-            //   new Order
-            //{
-            //    Id = 5,
-            //    No = "0005",
-            //    OrderName = "Monitor",
-            //    DocDate = DateTime.Today,
-            //    CustomerName = "Oming",
-            //    Amount = 6500
-            //},
-            //    new Order
-            //{
-            //    Id = 6,
-            //    No = "0006",
-            //    OrderName = "Monitor",
-            //    DocDate = DateTime.Today,
-            //    CustomerName = "Pedro",
-            //    Amount = 6500
-            //}
-        };
-    }
-}
