--- MVC_Activity01/MVC_Activity01/Views/Customer/CustomerView.cshtml
@@ -1,38 +1,0 @@
-﻿@*@model IEnumerable<MVC_Activity01.Models.CustomersDTO>
-
-<div class="jumbotron">
-
-    <h3>@ViewBag.Title</h3>
-    <div style="margin-left:86%; margin-bottom:30px;">
-        @Html.ActionLink("Add Customer", "Create", null, new { @class = "btn btn-success btn-md" })
-
-    </div>
-    <table style="width:100%" class="table-responsive">
-
-        <tr>
-            <th>Firstname</th>
-            <th>Middlename</th>
-            <th>Lastname</th>
-            <th>Birthday</th>
-            <th>Gender</th>
-            <th>Age</th>
-            <th>Address</th>
-            <th>Email</th>
-            <th>Action</th>
-        </tr>
-        @foreach (var item in Model)
-        {
-            <tr>
-                <td>@item.Firstname</td>
-                <td>@item.Middlename</td>
-                <td>@item.Lastname</td>
-                <td>@item.Birthday.ToShortDateString()</td>
-                <td>@item.Gender</td>
-                <td>@item.Age</td>
-                <td>@item.Address</td>
-                <td>@item.EmailAdress</td>
-                <td> @Html.ActionLink("Edit", "CustomerView", new { id = item.Id }, new { @class = "btn btn-primary btn-sm" })</td>
-            </tr>
-        }
-    </table>
-</div>*@
