+++ MVC_Activity01/MVC_Activity01/Views/Home/Index.cshtml
@@ -1,25 +1,1 @@
 ﻿
-<div class="jumbotron">
-    <h3>@ViewBag.Title</h3>
-    @Html.DropDownList("CustomerList", null, "Select Name")
-
-    @{
-        Html.RenderAction("OrderList", "Home");
-    }
-</div>
-
-@section scripts {
-    <script>
-        $("#CustomerList").change(function () {
-            var id = this.value;
-            $.ajax({
-                url: '@Url.Action("OrderList", "Home")',
-                data: { id: id },
-                type: 'POST',
-                success: function (result) {
-                    $("#ordList").html(result);
-                }
-            });
-        });
-    </script>
-}
