--- MVC_Activity01/MVC_Activity01/Views/Home/OrderDetails.cshtml
@@ -1,22 +1,0 @@
-﻿@model MVC_Activity01.Models.Order
-
-<h2>@ViewBag.Message</h2>
-
-<label>Order Number:@Model.No </label>
-<br />
-<br />
-<label>Order Date:@Model.DocDate.ToShortDateString() </label>
-<br />
-<br />
-@*<label>Customer Name:@Model.CustomerName</label>*@
-<br />
-<br />
-<label>Total Amount:@Model.Amount </label>
-<br />
-<br />
-<div>
-    <a href="/Home/"onMouseOver="this.style.color='#0F0'"
-   onMouseOut="this.style.color='#00F'" class="btn btn-primary">
-        Back
-    </a>
-</div>
