+++ MVC_Activity01/MVC_Activity01/Views/Shared/_Layout.cshtml
@@ -24,7 +24,7 @@
                     <li>@Html.ActionLink("About", "About", "Home")</li>
                     <li>@Html.ActionLink("Contact", "Contact", "Home")</li>
                     <li>@Html.ActionLink("Customer", "Index", "Customer")</li>
-                    <li>@Html.ActionLink("Orders", "Index", "Home")</li>
+                    <li>@Html.ActionLink("Orders", "Index", "Order")</li>
 
                 </ul>
             </div>
